added AI BarcodeParser (https://github.com/hprange/BarcodeParser)
===================================================================
...
and use the single one function parseBarcode() of the library, handling over the barcode string:

try {
        var barcode = document.getElementById("barcode").value,
            answer = parseBarcode(barcode);
        // handle the answer ...
} catch (e) {
    alert(e);
}
The function returns an object containing two elements:

codeName: a barcode type identifier (a simple string denoting the type of barcode) and
parsedCodeItems: an array of objects, each with four attributes:
ai: the application identifier
title: the title of the element, i.e. a short description
data: the contents, either a string, a number or a date
unit: the unit of measurement, a country code, a currency; denoted in ISO codes.
...
===================================================================

This could be used in a dataEdit.js file as follows:

function dataEdit(inStr, sAimID) {

 var vOut=inStr;
  try {
        var barcode = inStr;
        var answer = parseBarcode(barcode);
        vOut="";
        for (x=0; x<answer.parsedCodeItems.length; x++){
	   vOut+=answer.parsedCodeItems[x].ai + "("+ answer.parsedCodeItems[x].title + "): " + answer.parsedCodeItems[x].data + "\r\n";
        }
        return vOut;
  } catch (e) {
    return e; //alert(e);
  }